requirements.txt
aniso8601==9.0.1
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.25.0
numpy==1.20.2
pandas==1.2.4
pycparser==2.20
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.14
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
Werkzeug==1.0.1
zipp==3.4.1
